Add CLI for benchmarking tool
