+++ json_stream_rs_tokenizer/benchmark/__main__.py
@@ -1,4 +1,4 @@
-from . import main
+from .cli import main
 
 if __name__ == "__main__":
-    main()
+    exit(main())
